+++ sources/Google.Solutions.Mvvm/Binding/BindingContext.cs
