+++ sources/Google.Solutions.Mvvm/Binding/PropertiesView.resx
